refactor(portal-web): isolate rsbuild asset config
